Split unit and driver tests following ebpf-for-windows pattern
- ntosebpfext_unit: Keep only pure user-mode tests (process_invoke,
  process exit codes, process create and exit times, libbpf attach type
  names) that use usersim mock kernel APIs
- ntosebpfext_driver_test: New project with driver tests moved from
  ntosebpfext_unit (process_bpf_prog_run_test, process_resolve_account)
  that load real .sys drivers
- neteventebpfext_unit: Renamed to neteventebpfext_driver_test (all tests
  are kernel integration tests)
- Updated cicd.yml with renamed/new test jobs
- Updated codecov.yml after_n_builds count
- Updated docs/neteventebpfext.md references

Copilot-Session: 940d0edb-486d-4adf-a1a1-3f72fd7d4b66
