+++ .github/codecov.yml
@@ -22,9 +22,9 @@ codecov:
     #         add the number of comma-sepatated configurations to after_n_builds.
     # 'after_n_builds' is now set to the right count.
     #
-    after_n_builds: 2
+    after_n_builds: 3
 comment:
-    after_n_builds: 2
+    after_n_builds: 3
 coverage:
   status:
     project:
