+++ ntosebpfext.sln
@@ -36,7 +36,9 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "netevent_sim", "tests\netev
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ntosebpfext_unit", "tests\ntosebpfext\ntosebpfext_unit\ntosebpfext_unit.vcxproj", "{E619B985-44B3-4292-A585-5CD0C4315ED3}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "neteventebpfext_unit", "tests\neteventebpfext\neteventebpfext_unit\neteventebpfext_unit.vcxproj", "{32DA760D-1A43-4940-BECE-59AE0B817ED9}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "ntosebpfext_driver_test", "tests\ntosebpfext\ntosebpfext_driver_test\ntosebpfext_driver_test.vcxproj", "{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "neteventebpfext_driver_test", "tests\neteventebpfext\neteventebpfext_unit\neteventebpfext_unit.vcxproj", "{32DA760D-1A43-4940-BECE-59AE0B817ED9}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ntosebpfext", "ntosebpfext", "{ED8BFFE5-778D-4254-9800-6506F0C9CA5B}"
 EndProject
@@ -203,6 +205,14 @@ Global
 		{E619B985-44B3-4292-A585-5CD0C4315ED3}.Release|ARM64.Build.0 = Release|x64
 		{E619B985-44B3-4292-A585-5CD0C4315ED3}.Release|x64.ActiveCfg = Release|x64
 		{E619B985-44B3-4292-A585-5CD0C4315ED3}.Release|x64.Build.0 = Release|x64
+		{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}.Debug|ARM64.ActiveCfg = Debug|x64
+		{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}.Debug|ARM64.Build.0 = Debug|x64
+		{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}.Debug|x64.ActiveCfg = Debug|x64
+		{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}.Debug|x64.Build.0 = Debug|x64
+		{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}.Release|ARM64.ActiveCfg = Release|x64
+		{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}.Release|ARM64.Build.0 = Release|x64
+		{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}.Release|x64.ActiveCfg = Release|x64
+		{A1B2C3D4-E5F6-7890-ABCD-EF1234567890}.Release|x64.Build.0 = Release|x64
 		{32DA760D-1A43-4940-BECE-59AE0B817ED9}.Debug|ARM64.ActiveCfg = Debug|x64
 		{32DA760D-1A43-4940-BECE-59AE0B817ED9}.Debug|ARM64.Build.0 = Debug|x64
 		{32DA760D-1A43-4940-BECE-59AE0B817ED9}.Debug|x64.ActiveCfg = Debug|x64
@@ -336,6 +346,7 @@ Global
 		{630BB78F-6211-41D8-8E3A-096E22E169EF} = {E8184DFB-C7EA-4913-9F63-6E61D4B9CBB1}
 		{B42AB575-67A9-4BA1-8AA1-471AA3FD810C} = {1808952B-4865-4EDF-9393-8E52F60D4D20}
 		{E619B985-44B3-4292-A585-5CD0C4315ED3} = {E8184DFB-C7EA-4913-9F63-6E61D4B9CBB1}
+		{A1B2C3D4-E5F6-7890-ABCD-EF1234567890} = {E8184DFB-C7EA-4913-9F63-6E61D4B9CBB1}
 		{32DA760D-1A43-4940-BECE-59AE0B817ED9} = {1808952B-4865-4EDF-9393-8E52F60D4D20}
 		{ED8BFFE5-778D-4254-9800-6506F0C9CA5B} = {54A23059-C3A5-4527-9AA6-F8850930D838}
 		{C488132C-7C22-4A37-821A-D7656C485372} = {54A23059-C3A5-4527-9AA6-F8850930D838}
