+++ tests/ntosebpfext/ntosebpfext_unit/ntos_ebpfext_unit.cpp
@@ -15,17 +15,6 @@
 #include <bpf/bpf.h>
 #include <bpf/libbpf.h>
 #include <map>
-#pragma warning(push)
-#pragma warning(disable : 28182) // Dereferencing NULL pointer. 'Temp_value_#12076' contains the same NULL
-                                 //  value as 'new(1*144, nothrow)' did.
-#include <wil/resource.h>
-#pragma warning(pop)
-
-#define MAX_IMAGE_PATH_SIZE (1024)
-#define MAX_COMMAND_LINE_SIZE ((64 * 1024))
-#define ACCOUNT_NAME_SIZE (64)
-#define ACCOUNT_DOMAIN_SIZE (512)
-#define TEST_PROCESS_ID 1234
 
 struct _DEVICE_OBJECT* _ebpf_ext_driver_device_object;
 
@@ -34,71 +23,11 @@ CATCH_REGISTER_LISTENER(cxplat_passed_test_log)
 
 #pragma region process
 
-static std::atomic<uint32_t> process_event_count = 0;
-
-// Define process_info_t to match the structure in process_monitor.c
-typedef struct
-{
-    uint32_t process_id;
-    uint32_t parent_process_id;
-    uint32_t creating_process_id;
-    uint32_t creating_thread_id;
-    uint64_t creation_time;
-    uint64_t exit_time;
-    uint32_t process_exit_code;
-    uint8_t operation;
-    uint32_t token_sid_size;
-    uint8_t token_sid[TOKEN_SID_MAX_SIZE];
-} process_info_t;
-
-static int
-process_ringbuf_event_callback(void* ctx, void* data, size_t size)
-{
-    UNREFERENCED_PARAMETER(ctx);
-
-    if (size != sizeof(process_info_t)) {
-        std::cout << "Unexpected data size in ring buffer: " << size << " (expected " << sizeof(process_info_t) << ")"
-                  << std::endl;
-        return 0;
-    }
-
-    process_info_t* info = (process_info_t*)data;
-
-    if (info->process_id == TEST_PROCESS_ID) {
-        std::cout << "Ring buffer event received:" << std::endl;
-        std::cout << "  Process ID: " << info->process_id << std::endl;
-        std::cout << "  Parent Process ID: " << info->parent_process_id << std::endl;
-        std::cout << "  Creating Process ID: " << info->creating_process_id << std::endl;
-        std::cout << "  Creating Thread ID: " << info->creating_thread_id << std::endl;
-        std::cout << "  Creation Time: " << info->creation_time << std::endl;
-        std::cout << "  Exit Time: " << info->exit_time << std::endl;
-        std::cout << "  Exit Code: " << info->process_exit_code << std::endl;
-        std::cout << "  Operation: " << (info->operation == 0 ? "CREATE" : "DELETE") << std::endl;
-        std::cout << "  Token SID Size: " << info->token_sid_size << std::endl;
-        process_event_count++;
-    }
-
-    return 0;
-}
-
 typedef struct test_process_client_context_t
 {
     ntosebpfext_helper_base_client_context_t base;
     process_md_t process_context;
 } test_process_client_context_t;
-
-typedef struct test_process_notify_context
-{
-    EBPF_CONTEXT_HEADER;
-    process_md_t process_md;
-    void* process;
-    void* create_info;
-    UNICODE_STRING command_line;
-    UNICODE_STRING image_file_name;
-    UNICODE_STRING account_name;
-    UNICODE_STRING account_domain;
-    BOOLEAN account_lookup_done;
-} test_process_notify_context_t;
 
 _Must_inspect_result_ ebpf_result_t
 ntosebpfext_unit_invoke_process_program(
@@ -321,423 +250,6 @@ TEST_CASE("libbpf attach type names", "[ntosebpfext][libbpf]")
 
     REQUIRE(libbpf_attach_type_by_name(type_str, &attach_type) == 0);
     REQUIRE(attach_type == BPF_ATTACH_TYPE_PROCESS);
-}
-
-TEST_CASE("process_bpf_prog_run_test", "[ntosebpfext]")
-{
-    // Load and start ntosebpfext extension driver.
-    driver_service ntosebpfext_driver;
-    REQUIRE(
-        ntosebpfext_driver.create(L"ntosebpfext", driver_service::get_driver_path("ntosebpfext.sys").c_str()) == true);
-    REQUIRE(ntosebpfext_driver.start() == true);
-    auto cleanup_driver = wil::scope_exit([&]() {
-        ntosebpfext_driver.stop();
-        ntosebpfext_driver.unload();
-    });
-
-    // Load the process monitor BPF program.
-    struct bpf_object* object = bpf_object__open("process_monitor.sys");
-    REQUIRE(object != nullptr);
-    auto cleanup_object = wil::scope_exit([&]() {
-        if (object != nullptr) {
-            bpf_object__close(object);
-        }
-    });
-
-    int res = bpf_object__load(object);
-    REQUIRE(res == 0);
-
-    // Find the process monitor BPF program.
-    bpf_program* process_monitor = bpf_object__find_program_by_name(object, "ProcessMonitor");
-    REQUIRE(process_monitor != nullptr);
-
-    // Initialize structures required for bpf_prog_test_run_opts
-    bpf_test_run_opts bpf_opts = {0};
-    test_process_notify_context_t process_ctx_in = {0};
-    test_process_notify_context_t process_ctx_out = {0};
-
-    fd_t process_program_fd = bpf_program__fd(process_monitor);
-    REQUIRE(process_program_fd != ebpf_fd_invalid);
-
-    // Prepare test process data
-    std::wstring command_line = L"notepad.exe test.txt";
-    std::wstring image_path = L"C:\\Windows\\System32\\notepad.exe";
-    std::wstring account_name = L"SYSTEM";
-    std::wstring account_domain = L"NT AUTHORITY";
-
-    process_ctx_in.process_md.process_id = TEST_PROCESS_ID;
-    process_ctx_in.process_md.parent_process_id = 4567;
-    process_ctx_in.process_md.creating_process_id = 8910;
-    process_ctx_in.process_md.creating_thread_id = 1112;
-    process_ctx_in.process_md.operation = PROCESS_OPERATION_CREATE;
-    process_ctx_in.process_md.process_exit_code = 0;
-    process_ctx_in.process_md.creation_time = 123456789;
-    process_ctx_in.process_md.exit_time = 0;
-
-    // Set up a test SID (well-known Local System SID: S-1-5-18).
-    uint8_t test_sid[] = {
-        0x01, // Revision
-        0x01, // SubAuthorityCount
-        0x00,
-        0x00,
-        0x00,
-        0x00,
-        0x00,
-        0x05, // IdentifierAuthority (NT Authority)
-        0x12,
-        0x00,
-        0x00,
-        0x00 // SubAuthority[0] = 18 (SECURITY_LOCAL_SYSTEM_RID)
-    };
-    process_ctx_in.process_md.token_sid_size = sizeof(test_sid);
-    memcpy(process_ctx_in.process_md.token_sid, test_sid, sizeof(test_sid));
-
-    // Set up UNICODE_STRING structures with only Length fields (Buffer will be NULL)
-    // The actual data will be passed in data_in buffer
-    process_ctx_in.command_line.Length = static_cast<USHORT>(command_line.length() * sizeof(wchar_t));
-    process_ctx_in.command_line.MaximumLength = process_ctx_in.command_line.Length;
-    process_ctx_in.command_line.Buffer = NULL;
-
-    process_ctx_in.image_file_name.Length = static_cast<USHORT>(image_path.length() * sizeof(wchar_t));
-    process_ctx_in.image_file_name.MaximumLength = process_ctx_in.image_file_name.Length;
-    process_ctx_in.image_file_name.Buffer = NULL;
-
-    process_ctx_in.account_name.Length = static_cast<USHORT>(account_name.length() * sizeof(wchar_t));
-    process_ctx_in.account_name.MaximumLength = process_ctx_in.account_name.Length;
-    process_ctx_in.account_name.Buffer = NULL;
-
-    process_ctx_in.account_domain.Length = static_cast<USHORT>(account_domain.length() * sizeof(wchar_t));
-    process_ctx_in.account_domain.MaximumLength = process_ctx_in.account_domain.Length;
-    process_ctx_in.account_domain.Buffer = NULL;
-
-    // Manually set command_start and command_end pointers to NULL (will be set by context_create)
-    process_ctx_in.process_md.command_start = NULL;
-    process_ctx_in.process_md.command_end = NULL;
-
-    // Pack command_line, image_file_name, account_name, and account_domain data into a single buffer for data_in
-    size_t total_data_size = static_cast<size_t>(process_ctx_in.command_line.Length) +
-                             static_cast<size_t>(process_ctx_in.image_file_name.Length) +
-                             static_cast<size_t>(process_ctx_in.account_name.Length) +
-                             static_cast<size_t>(process_ctx_in.account_domain.Length);
-    std::vector<uint8_t> packed_data(total_data_size);
-
-    size_t offset = 0;
-    memcpy(packed_data.data() + offset, command_line.c_str(), process_ctx_in.command_line.Length);
-    offset += process_ctx_in.command_line.Length;
-    memcpy(packed_data.data() + offset, image_path.c_str(), process_ctx_in.image_file_name.Length);
-    offset += process_ctx_in.image_file_name.Length;
-    memcpy(packed_data.data() + offset, account_name.c_str(), process_ctx_in.account_name.Length);
-    offset += process_ctx_in.account_name.Length;
-    memcpy(packed_data.data() + offset, account_domain.c_str(), process_ctx_in.account_domain.Length);
-
-    // Set up ring buffer consumer before running the test
-    bpf_map* process_ringbuf_map = bpf_object__find_map_by_name(object, "process_ringbuf");
-    REQUIRE(process_ringbuf_map != nullptr);
-    int process_ringbuf_fd = bpf_map__fd(process_ringbuf_map);
-    REQUIRE(process_ringbuf_fd != ebpf_fd_invalid);
-
-    ring_buffer* process_ring_buffer =
-        ring_buffer__new(process_ringbuf_fd, process_ringbuf_event_callback, nullptr, nullptr);
-    REQUIRE(process_ring_buffer != nullptr);
-    auto cleanup_ring_buffer = wil::scope_exit([&]() {
-        if (process_ring_buffer != nullptr) {
-            ring_buffer__free(process_ring_buffer);
-        }
-    });
-
-    // Prepare buffer for data_out
-    std::vector<uint8_t> data_out_buffer(total_data_size);
-
-    // Prepare bpf_opts
-    bpf_opts.repeat = 1;
-    bpf_opts.ctx_in = &process_ctx_in;
-    bpf_opts.ctx_size_in = sizeof(process_ctx_in);
-    bpf_opts.ctx_out = &process_ctx_out;
-    bpf_opts.ctx_size_out = sizeof(process_ctx_out);
-    bpf_opts.data_in = packed_data.data();
-    bpf_opts.data_size_in = static_cast<uint32_t>(total_data_size);
-    bpf_opts.data_out = data_out_buffer.data();
-    bpf_opts.data_size_out = static_cast<uint32_t>(total_data_size);
-
-    // Execute the program - expect success
-    REQUIRE(bpf_prog_test_run_opts(process_program_fd, &bpf_opts) == 0);
-
-    // Validate the output context and data
-    REQUIRE(bpf_opts.ctx_size_out == sizeof(process_ctx_out));
-    REQUIRE(bpf_opts.data_size_out == bpf_opts.data_size_in);
-    REQUIRE(process_ctx_out.process_md.process_id == process_ctx_in.process_md.process_id);
-    REQUIRE(process_ctx_out.process_md.parent_process_id == process_ctx_in.process_md.parent_process_id);
-    REQUIRE(process_ctx_out.process_md.operation == process_ctx_in.process_md.operation);
-    REQUIRE(process_ctx_out.process_md.token_sid_size == process_ctx_in.process_md.token_sid_size);
-    REQUIRE(
-        memcmp(
-            process_ctx_out.process_md.token_sid,
-            process_ctx_in.process_md.token_sid,
-            process_ctx_in.process_md.token_sid_size) == 0);
-
-    // Validate that exactly one event was written to the ring buffer
-    uint32_t event_count_before = process_event_count;
-    REQUIRE(ring_buffer__poll(process_ring_buffer, 5000) > 0);
-    REQUIRE(process_event_count == event_count_before + 1);
-
-    // Validate LRU_HASH maps: process_map and command_map
-    bpf_map* process_map = bpf_object__find_map_by_name(object, "process_map");
-    REQUIRE(process_map != nullptr);
-    int process_map_fd = bpf_map__fd(process_map);
-    REQUIRE(process_map_fd != ebpf_fd_invalid);
-
-    bpf_map* command_map = bpf_object__find_map_by_name(object, "command_map");
-    REQUIRE(command_map != nullptr);
-    int command_map_fd = bpf_map__fd(command_map);
-    REQUIRE(command_map_fd != ebpf_fd_invalid);
-
-    // Lookup the process_id in process_map to verify image path was stored
-    uint32_t lookup_key = (uint32_t)process_ctx_in.process_md.process_id;
-    std::vector<wchar_t> image_path_from_map(MAX_IMAGE_PATH_SIZE / sizeof(wchar_t));
-    int result_process = bpf_map_lookup_elem(process_map_fd, &lookup_key, image_path_from_map.data());
-    REQUIRE(result_process == 0);
-    REQUIRE(wcscmp(image_path_from_map.data(), image_path.c_str()) == 0);
-
-    // Lookup the process_id in command_map to verify command line was stored
-    std::vector<wchar_t> command_line_from_map(MAX_COMMAND_LINE_SIZE / sizeof(wchar_t));
-    int result_command = bpf_map_lookup_elem(command_map_fd, &lookup_key, command_line_from_map.data());
-    REQUIRE(result_command == 0);
-    REQUIRE(wcscmp(command_line_from_map.data(), command_line.c_str()) == 0);
-
-    // Validate LRU_HASH maps: account_name_map and account_domain_map
-    bpf_map* account_name_map = bpf_object__find_map_by_name(object, "account_name_map");
-    REQUIRE(account_name_map != nullptr);
-    int account_name_map_fd = bpf_map__fd(account_name_map);
-    REQUIRE(account_name_map_fd != ebpf_fd_invalid);
-
-    bpf_map* account_domain_map = bpf_object__find_map_by_name(object, "account_domain_map");
-    REQUIRE(account_domain_map != nullptr);
-    int account_domain_map_fd = bpf_map__fd(account_domain_map);
-    REQUIRE(account_domain_map_fd != ebpf_fd_invalid);
-
-    // Lookup the process_id in account_name_map to verify account name was stored
-    std::vector<wchar_t> account_name_from_map(ACCOUNT_NAME_SIZE / sizeof(wchar_t));
-    int result_account_name = bpf_map_lookup_elem(account_name_map_fd, &lookup_key, account_name_from_map.data());
-    REQUIRE(result_account_name == 0);
-    REQUIRE(wcscmp(account_name_from_map.data(), account_name.c_str()) == 0);
-
-    // Lookup the process_id in account_domain_map to verify account domain was stored
-    std::vector<wchar_t> account_domain_from_map(ACCOUNT_DOMAIN_SIZE / sizeof(wchar_t));
-    int result_account_domain = bpf_map_lookup_elem(account_domain_map_fd, &lookup_key, account_domain_from_map.data());
-    REQUIRE(result_account_domain == 0);
-    REQUIRE(wcscmp(account_domain_from_map.data(), account_domain.c_str()) == 0);
-
-    // Test negative cases
-
-    // Context smaller than process_notify_context_t must be rejected
-    unsigned char smaller_ctx[sizeof(process_ctx_in) - 1] = {0};
-    bpf_opts.ctx_in = &smaller_ctx;
-    bpf_opts.ctx_size_in = sizeof(smaller_ctx);
-    REQUIRE(bpf_prog_test_run_opts(process_program_fd, &bpf_opts) != 0);
-
-    // NULL context should be rejected
-    bpf_opts.ctx_in = nullptr;
-    bpf_opts.ctx_size_in = 0;
-    REQUIRE(bpf_prog_test_run_opts(process_program_fd, &bpf_opts) != 0);
-
-    // NULL data_in with valid context should be rejected
-    bpf_opts.ctx_in = &process_ctx_in;
-    bpf_opts.ctx_size_in = sizeof(process_ctx_in);
-    bpf_opts.data_in = nullptr;
-    bpf_opts.data_size_in = 0;
-    REQUIRE(bpf_prog_test_run_opts(process_program_fd, &bpf_opts) != 0);
-
-    // Insufficient data_size_in should be rejected
-    // Set data_size_in to 1 byte less than required (command_line.Length + image_file_name.Length)
-    bpf_opts.data_in = packed_data.data();
-    bpf_opts.data_size_in = static_cast<uint32_t>(total_data_size - 1);
-    REQUIRE(bpf_prog_test_run_opts(process_program_fd, &bpf_opts) != 0);
-}
-
-TEST_CASE("process_resolve_account", "[ntosebpfext]")
-{
-    // Exercise _ebpf_process_resolve_account via bpf_prog_test_run.
-    // context_create no longer forces account_lookup_done = TRUE, so the BPF program's call
-    // to bpf_process_get_account_name triggers the lazy resolution using token_sid.
-    driver_service ntosebpfext_driver;
-    REQUIRE(
-        ntosebpfext_driver.create(L"ntosebpfext", driver_service::get_driver_path("ntosebpfext.sys").c_str()) == true);
-    REQUIRE(ntosebpfext_driver.start() == true);
-    auto cleanup_driver = wil::scope_exit([&]() {
-        ntosebpfext_driver.stop();
-        ntosebpfext_driver.unload();
-    });
-
-    struct bpf_object* object = bpf_object__open("process_monitor.sys");
-    REQUIRE(object != nullptr);
-    auto cleanup_object = wil::scope_exit([&]() { bpf_object__close(object); });
-
-    REQUIRE(bpf_object__load(object) == 0);
-
-    bpf_program* process_monitor = bpf_object__find_program_by_name(object, "ProcessMonitor");
-    REQUIRE(process_monitor != nullptr);
-
-    fd_t process_program_fd = bpf_program__fd(process_monitor);
-    REQUIRE(process_program_fd != ebpf_fd_invalid);
-
-    bpf_map* process_ringbuf_map = bpf_object__find_map_by_name(object, "process_ringbuf");
-    REQUIRE(process_ringbuf_map != nullptr);
-    ring_buffer* process_ring_buffer =
-        ring_buffer__new(bpf_map__fd(process_ringbuf_map), process_ringbuf_event_callback, nullptr, nullptr);
-    REQUIRE(process_ring_buffer != nullptr);
-    auto cleanup_ring_buffer = wil::scope_exit([&]() { ring_buffer__free(process_ring_buffer); });
-
-    bpf_map* account_name_map = bpf_object__find_map_by_name(object, "account_name_map");
-    REQUIRE(account_name_map != nullptr);
-    int account_name_map_fd = bpf_map__fd(account_name_map);
-    REQUIRE(account_name_map_fd != ebpf_fd_invalid);
-
-    bpf_map* account_domain_map = bpf_object__find_map_by_name(object, "account_domain_map");
-    REQUIRE(account_domain_map != nullptr);
-    int account_domain_map_fd = bpf_map__fd(account_domain_map);
-    REQUIRE(account_domain_map_fd != ebpf_fd_invalid);
-
-    std::wstring command_line = L"notepad.exe test.txt";
-    std::wstring image_path = L"C:\\Windows\\System32\\notepad.exe";
-    uint8_t test_sid[] = {
-        0x01, // Revision
-        0x01, // SubAuthorityCount
-        0x00,
-        0x00,
-        0x00,
-        0x00,
-        0x00,
-        0x05, // IdentifierAuthority (NT Authority)
-        0x12,
-        0x00,
-        0x00,
-        0x00 // SubAuthority[0] = 18 (SECURITY_LOCAL_SYSTEM_RID)
-    };
-
-    // --- Scenario 1: Heap-fallback path ---
-    // Provide a valid SID but tiny account buffers (MaximumLength = 2) so SecLookupAccountSid
-    // returns STATUS_BUFFER_TOO_SMALL on the first try, forcing heap allocation and retry.
-    {
-        bpf_test_run_opts bpf_opts = {0};
-        test_process_notify_context_t process_ctx_in = {0};
-        test_process_notify_context_t process_ctx_out = {0};
-
-        process_ctx_in.process_md.process_id = TEST_PROCESS_ID;
-        process_ctx_in.process_md.operation = PROCESS_OPERATION_CREATE;
-        process_ctx_in.process_md.token_sid_size = sizeof(test_sid);
-        memcpy(process_ctx_in.process_md.token_sid, test_sid, sizeof(test_sid));
-
-        process_ctx_in.command_line.Length = static_cast<USHORT>(command_line.length() * sizeof(wchar_t));
-        process_ctx_in.command_line.MaximumLength = process_ctx_in.command_line.Length;
-        process_ctx_in.image_file_name.Length = static_cast<USHORT>(image_path.length() * sizeof(wchar_t));
-        process_ctx_in.image_file_name.MaximumLength = process_ctx_in.image_file_name.Length;
-
-        wchar_t dummy = L'X';
-        process_ctx_in.account_name.Length = sizeof(dummy);
-        process_ctx_in.account_name.MaximumLength = sizeof(dummy);
-        process_ctx_in.account_domain.Length = sizeof(dummy);
-        process_ctx_in.account_domain.MaximumLength = sizeof(dummy);
-
-        size_t total_data_size = static_cast<size_t>(process_ctx_in.command_line.Length) +
-                                 static_cast<size_t>(process_ctx_in.image_file_name.Length) +
-                                 static_cast<size_t>(process_ctx_in.account_name.Length) +
-                                 static_cast<size_t>(process_ctx_in.account_domain.Length);
-        std::vector<uint8_t> packed_data(total_data_size);
-
-        size_t offset = 0;
-        memcpy(packed_data.data() + offset, command_line.c_str(), process_ctx_in.command_line.Length);
-        offset += process_ctx_in.command_line.Length;
-        memcpy(packed_data.data() + offset, image_path.c_str(), process_ctx_in.image_file_name.Length);
-        offset += process_ctx_in.image_file_name.Length;
-        memcpy(packed_data.data() + offset, &dummy, process_ctx_in.account_name.Length);
-        offset += process_ctx_in.account_name.Length;
-        memcpy(packed_data.data() + offset, &dummy, process_ctx_in.account_domain.Length);
-
-        std::vector<uint8_t> data_out_buffer(1024);
-
-        bpf_opts.repeat = 1;
-        bpf_opts.ctx_in = &process_ctx_in;
-        bpf_opts.ctx_size_in = sizeof(process_ctx_in);
-        bpf_opts.ctx_out = &process_ctx_out;
-        bpf_opts.ctx_size_out = sizeof(process_ctx_out);
-        bpf_opts.data_in = packed_data.data();
-        bpf_opts.data_size_in = static_cast<uint32_t>(total_data_size);
-        bpf_opts.data_out = data_out_buffer.data();
-        bpf_opts.data_size_out = static_cast<uint32_t>(data_out_buffer.size());
-
-        REQUIRE(bpf_prog_test_run_opts(process_program_fd, &bpf_opts) == 0);
-
-        uint32_t event_count_before_s1 = process_event_count;
-        REQUIRE(ring_buffer__poll(process_ring_buffer, 5000) > 0);
-        REQUIRE(process_event_count == event_count_before_s1 + 1);
-
-        uint32_t lookup_key = TEST_PROCESS_ID;
-        std::vector<wchar_t> account_name_from_map(ACCOUNT_NAME_SIZE / sizeof(wchar_t), L'\0');
-        REQUIRE(bpf_map_lookup_elem(account_name_map_fd, &lookup_key, account_name_from_map.data()) == 0);
-        REQUIRE(account_name_from_map[0] != L'\0');
-
-        std::vector<wchar_t> account_domain_from_map(ACCOUNT_DOMAIN_SIZE / sizeof(wchar_t), L'\0');
-        REQUIRE(bpf_map_lookup_elem(account_domain_map_fd, &lookup_key, account_domain_from_map.data()) == 0);
-        REQUIRE(account_domain_from_map[0] != L'\0');
-
-        // Delete the map entries so scenario 2 starts clean.
-        bpf_map_delete_elem(account_name_map_fd, &lookup_key);
-        bpf_map_delete_elem(account_domain_map_fd, &lookup_key);
-    }
-
-    // --- Scenario 2: Zero SID ---
-    // token_sid_size == 0 causes _ebpf_process_resolve_account to return failure,
-    // so the helpers return -ENOENT and the BPF program skips account map writes.
-    {
-        bpf_test_run_opts bpf_opts = {0};
-        test_process_notify_context_t process_ctx_in = {0};
-        test_process_notify_context_t process_ctx_out = {0};
-
-        process_ctx_in.process_md.process_id = TEST_PROCESS_ID;
-        process_ctx_in.process_md.operation = PROCESS_OPERATION_CREATE;
-        process_ctx_in.process_md.token_sid_size = 0;
-
-        process_ctx_in.command_line.Length = static_cast<USHORT>(command_line.length() * sizeof(wchar_t));
-        process_ctx_in.command_line.MaximumLength = process_ctx_in.command_line.Length;
-        process_ctx_in.image_file_name.Length = static_cast<USHORT>(image_path.length() * sizeof(wchar_t));
-        process_ctx_in.image_file_name.MaximumLength = process_ctx_in.image_file_name.Length;
-
-        size_t total_data_size = static_cast<size_t>(process_ctx_in.command_line.Length) +
-                                 static_cast<size_t>(process_ctx_in.image_file_name.Length);
-        std::vector<uint8_t> packed_data(total_data_size);
-
-        size_t offset = 0;
-        memcpy(packed_data.data() + offset, command_line.c_str(), process_ctx_in.command_line.Length);
-        offset += process_ctx_in.command_line.Length;
-        memcpy(packed_data.data() + offset, image_path.c_str(), process_ctx_in.image_file_name.Length);
-
-        std::vector<uint8_t> data_out_buffer(total_data_size);
-
-        bpf_opts.repeat = 1;
-        bpf_opts.ctx_in = &process_ctx_in;
-        bpf_opts.ctx_size_in = sizeof(process_ctx_in);
-        bpf_opts.ctx_out = &process_ctx_out;
-        bpf_opts.ctx_size_out = sizeof(process_ctx_out);
-        bpf_opts.data_in = packed_data.data();
-        bpf_opts.data_size_in = static_cast<uint32_t>(total_data_size);
-        bpf_opts.data_out = data_out_buffer.data();
-        bpf_opts.data_size_out = static_cast<uint32_t>(total_data_size);
-
-        REQUIRE(bpf_prog_test_run_opts(process_program_fd, &bpf_opts) == 0);
-
-        uint32_t event_count_before_s2 = process_event_count;
-        REQUIRE(ring_buffer__poll(process_ring_buffer, 5000) > 0);
-        REQUIRE(process_event_count == event_count_before_s2 + 1);
-
-        REQUIRE(process_ctx_out.process_md.token_sid_size == 0);
-
-        uint32_t lookup_key = TEST_PROCESS_ID;
-        std::vector<wchar_t> account_name_from_map(ACCOUNT_NAME_SIZE / sizeof(wchar_t));
-        REQUIRE(bpf_map_lookup_elem(account_name_map_fd, &lookup_key, account_name_from_map.data()) != 0);
-
-        std::vector<wchar_t> account_domain_from_map(ACCOUNT_DOMAIN_SIZE / sizeof(wchar_t));
-        REQUIRE(bpf_map_lookup_elem(account_domain_map_fd, &lookup_key, account_domain_from_map.data()) != 0);
-    }
 }
 
 #pragma endregion process
